docs: :memo: Update README by removing unnecessary information
This commit updates the README file by removing unnecessary information that is not relevant to users and contributors of the project. A thorough review of the content is conducted, and sections that are not pertinent or could cause confusion are eliminated. The goal is to provide a more concise README focused on key information to enhance understanding and usage of the project.